requirements.txt
et-xmlfile==1.1.0
nulltype==2.3.1
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
plaid-python==18.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4
urllib3==2.1.0
